requirements.txt
fastapi
uvicorn
redis
sqlalchemy
psycopg2-binary
python-dotenv
hashids
alembic
minio
awscli
